+++ web/src/main/java/com/tripper/dto/request/budget/UpdateCategoryAmountDto.java
@@ -1,4 +1,4 @@
-package com.tripper.dto.request;
+package com.tripper.dto.request.budget;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,7 +8,7 @@ import javax.validation.constraints.Positive;
 
 @Getter @Setter
 @AllArgsConstructor
-public class CreateOrUpdateBudgetDto {
+public class UpdateCategoryAmountDto {
     @Positive(message = "예산은 양수만 가능합니다.")
-    private int totalAmount;
+    private int amount;
 }
