+++ web/src/main/java/com/tripper/service/BudgetService.java
@@ -4,11 +4,11 @@ import com.tripper.domain.budget.Budget;
 import com.tripper.domain.budget.Category;
 import com.tripper.domain.budget.Item;
 import com.tripper.domain.trip.Trip;
-import com.tripper.dto.request.*;
-import com.tripper.dto.response.GetBudgetDto;
-import com.tripper.dto.response.GetBudgetItemDto;
-import com.tripper.dto.response.GetBudgetItemsDto;
-import com.tripper.dto.response.GetCategoryDto;
+import com.tripper.dto.request.budget.*;
+import com.tripper.dto.response.budget.GetBudgetDto;
+import com.tripper.dto.response.budget.GetBudgetItemDto;
+import com.tripper.dto.response.budget.GetBudgetItemsDto;
+import com.tripper.dto.response.budget.GetCategoryDto;
 import com.tripper.repository.BudgetRepository;
 import com.tripper.repository.TripRepository;
 import lombok.RequiredArgsConstructor;
