+++ web/src/main/java/com/tripper/service/TripService.java
@@ -2,15 +2,15 @@ package com.tripper.service;
 
 import com.tripper.domain.trip.Schedule;
 import com.tripper.domain.trip.Trip;
-import com.tripper.dto.request.CreateScheduleDto;
+import com.tripper.dto.request.trip.CreateScheduleDto;
 import com.tripper.domain.user.User;
-import com.tripper.dto.request.CreateTripDto;
-import com.tripper.dto.request.UpdateScheduleDto;
-import com.tripper.dto.request.UpdateTripDto;
-import com.tripper.dto.response.GetScheduleDto;
-import com.tripper.dto.response.GetScheduleListDto;
-import com.tripper.dto.response.GetTripDto;
-import com.tripper.dto.response.GetTripListDto;
+import com.tripper.dto.request.trip.CreateTripDto;
+import com.tripper.dto.request.trip.UpdateScheduleDto;
+import com.tripper.dto.request.trip.UpdateTripDto;
+import com.tripper.dto.response.trip.GetScheduleDto;
+import com.tripper.dto.response.trip.GetScheduleListDto;
+import com.tripper.dto.response.trip.GetTripDto;
+import com.tripper.dto.response.trip.GetTripListDto;
 import com.tripper.repository.TripRepository;
 import com.tripper.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
